requirements.txt
affine==2.4.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.6
decorator==5.1.1
ephem==4.1.5
exceptiongroup==1.2.0
executing==1.2.0
fastapi==0.110.2
flatbuffers==24.3.7
fonttools==4.50.0
h11==0.14.0
h5py==3.10.0
idna==3.6
importlib-metadata==6.0.0
importlib_resources==6.3.1
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.3
packaging==23.0
pandas==2.2.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
psutil==5.9.4
pure-eval==0.2.2
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.14.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==305
PyYAML==6.0.1
pyzmq==25.0.0
rasterio==1.3.9
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
stack-data==0.6.2
starlette==0.37.2
tornado==6.2
tppss==0.2
traitlets==5.9.0
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.6
zipp==3.14.0
